[FE] feat: game manual game 페이지에 추가
